Fixed some methods in Sql Class and fixed test method
